requirements.txt
pandas
numpy
requests
python-dotenv
xgboost
scikit-learn
fastapi
uvicorn
pydantic
pytz
joblib
